Extracted Stage6Planner from FirstJoinTask.PlanStage6 (debt-paydown PR 8)
* Added Stage6Planner collaborator owning the four cross-file planning phases (multi-charge consensus, consensus RTs, calibration refit, reconciliation planning)
* Each phase routes its diagnostic dump through ctx.Diagnostics, preserving dump call order
* Reduced PlanStage6 to a sequencer: plan, write reconciliation envelopes, publish byproducts
* Pure code motion; Stellar regression modes 1 and 2 byte-identical

See ai/todos/active/TODO-20260619_ospreysharp_debt_paydown_pr8.md
